fixed MAJOR bug that's been around for a while: sandbox/demo mode can allow deleting audio items from the db but should not allow actually deleting the A18 content, since that's still shared in dropbox, even in sandbox/demo mode

+++ acm/branches/acm_1_1_0/acm/src/main/java/org/literacybridge/acm/gui/dialogs/AudioItemContextMenuDialog.java
@@ -102,7 +102,9 @@ public class AudioItemContextMenuDialog extends JDialog implements WindowListene
 						for (AudioItem a : selectedAudioItems) {
 							try {
 								a.destroy();
-								Configuration.getRepository().delete(a);
+								// it's okay to delete from DB but cannot delete the .a18 file since that's in the shared (dropbox) repository
+								if (!ControlAccess.isACMReadOnly() && !ControlAccess.isSandbox())
+									Configuration.getRepository().delete(a);
 							} catch (Exception e) {
 								LOG.log(Level.WARNING, "Unable to delete audioitem id=" + a.getUuid(), e);
 							}
